requirements.txt
streamlit
openai
python-dotenv
requests
pyarrow